requirements.txt
azure-mgmt-security==1.0.0
azure-identity==1.9.0
